TableGen: Split RuntimeLibcallsEmitter into separate utility header
This information will be needed in more emitters, so start factoring
it to be more reusable.
